scripts/evaluate.py
"""Evaluation entry point for ASR checkpoints.

This public version is a scaffold intended for repository release.
Replace placeholder logic with the exact evaluation workflow used in the study.
"""

from __future__ import annotations

import argparse

import yaml


def parse_args() -> argparse.Namespace:
    parser = argparse.ArgumentParser()
    parser.add_argument("--config", required=True)
    parser.add_argument("--checkpoint", required=False, default=None)
    return parser.parse_args()


def main() -> None:
    args = parse_args()
    with open(args.config, "r", encoding="utf-8") as f:
        config = yaml.safe_load(f)

    checkpoint = args.checkpoint or config.get("checkpoint_path")
    print("Loaded evaluation configuration:")
    for key, value in config.items():
        print(f"  {key}: {value}")
    print(f"Checkpoint: {checkpoint}")

    raise NotImplementedError(
        "Insert the study-specific evaluation workflow here, including WER/CER reporting."
    )


if __name__ == "__main__":
    main()
