requirements.txt
torch==2.2.2
torchaudio==2.2.2
transformers==4.46.3
datasets==3.1.0
evaluate==0.4.3
jiwer==3.0.5
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
librosa==0.10.2.post1
soundfile==0.12.1
accelerate==1.0.1
PyYAML==6.0.2
tqdm==4.66.5
